cp: drop printf/scanf specifiers newlib-nano cannot parse

+++ common/eh_common/src/eh_check.c
@@ -31,6 +31,6 @@ void eh_check_failed_warn(esp_err_t rc, const char *file, int line,
 void eh_check_narrow_warn(long long val, const char *what, const char *function)
 {
     ESP_LOGW("EH_ASSIGN_NARROW",
-             "%s: value %lld does not fit %s — truncated on assignment",
-             function, val, what);
+             "%s: value %ld does not fit %s — truncated on assignment",
+             function, (long)val, what);
 }
